# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GabFlow API
  version: 0.4.0
  description: API principal para gestão de gabinetes municipais.
servers:
  - url: https://api.gabflow.example/api/v1
security:
  - cookieAuth: []
paths:
...
  /privacidade/solicitacoes/{solicitacaoId}:
    patch:
      summary: Atualiza validação, responsável, estado e resolução
      parameters:
        - in: path
          name: solicitacaoId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Solicitação atualizada e auditada
  /privacidade/solicitacoes/{solicitacaoId}/exportar:
    post:
      summary: Exporta pacote JSON após validação de identidade
      parameters:
        - in: path
          name: solicitacaoId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Pacote de dados do titular
        "422":
          description: Identidade não validada ou tipo incompatível
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
...